Fix a debug print statement that was longer than 80 characters
Change gpt_entry **pgpt_pte in print_part_efi() and get_partition_info_efi() to a single pointer and then passing the address of the pointer instead of a double pointer as an arguement to function calls. This was done because the existing code assigned the value of NULL to **pgpt_pte and then dereferenced it in is_gpt_valid() before doing any kind of memory allocation. Making it a single pointer and then passing it's address ensures that dereferencing it is safe.

Change-Id: I1f7a9e6da01724c0bcb29d851f0a0e33ecf8f8a6

Change gpt_entry **pgpt_pte to a single pointer

Change-Id: I31ea308d74e9851f1f0ff58d3aee84e01d337cd3

Review URL: http://codereview.chromium.org/3019022
Patch from Christopher Herrington <cherrint@codeaurora.org>.