Add generation of u-boot-cflags.mk
There are libraries (so far only vboot_reference) that will be linked with
u-boot.  These libraries have to be compiled with compatible compiler flags
so that linked code can be executed.  That is, there should be no mismatch
calling convention or errors with similar nature between u-boot and the
libraries.

This CL exports the compiler flags of u-boot, and any library that will be
linked with u-boot should use flags that are compatible with the exported
flags.

BUG=None
TEST=make u-boot-cflags.mk successfully

Review URL: http://codereview.chromium.org/6304022

Change-Id: Ib7afadc81791470350427ce7d913a5d48a32b7b0