Enable SPI flash environment as a dev option for Chrome OS
Enables a saved environment as the default for developer config. For Tegra2
this is an area at the end of SPI flash on Tegra2 boards.

BUG=chromiumos-11119
TEST=emerge-tegra2_kaen, flash and test, to check no saved environment
is active (this is chromeos_tegra2_kaen_recovery_config)
Also tested in the same way on seaboard, with J9 (UART ENABLE JMPR) in the
middle position (GPIO CTRL).

tegra2_kaen_config: make ..., flash and test environment
commands to check that it is accessible here also.

chromos_tegra2_kaen_developer_config: make ..., flash and test
environment to make sure saved environment is active.

Change-Id: Ibe78b29049386e116931dd3682190904c27cc7fe

Review URL: http://codereview.chromium.org/6298010