Create flasher configs for all Tegra2 boards.
The flasher configs are used to write a ChromeOS firmware payload to the SPI device on a Tegra2 board.  They are stripped down versions of U-Boot that expect a firmware image to be prepended to them.  They are written into RAM using the NVIDIA nvflash tool.

Change-Id: Ic8b713eed25707640abee90f99249fdd304a4b10

BUG=chromium-os:11981
TEST=emerge chromeos-u-boot-next for all boards, verify that a flasher image is build and use write_tegra_bios to run the image on each board out of RAM.  Verify that the image can in fact write to the SPI flash.

Review URL: http://codereview.chromium.org/6519004