Remove the board config data section.
It can be the first part of the bootstub as long as the bootstub is
the first section in the firmware blob.  And it needs to be the
first part fo the bootstub because of the way the Tegra tools
sign the bootstub.

Change-Id: I72fad3491ea84c4c245694c1ee742cd785bc7985

BUG=chromium-os:11981
TEST=emerge-tegra2_seaboard -av chromeos-u-boot-next

This CL depends on: http://codereview.chromium.org/6517012/

Review URL: http://codereview.chromium.org/6484027