Intialize aebl build target from kaen.
PTAL. This gets the build working for aebl but additional changes are required for u-boot to work properly.
Keyboard scan matrix comes from seaboard. UART configuration comes from kaen.

The following change is needed to work for the board to come out of reset:
http://codereview.chromium.org/6523018/

With this patched in, the serial console is available and it will start loading the kernel but does not make it to X.

Change-Id: I8460b04096d64f3863ea10d8c2f071f60c3aae34

BUG=None
TEST=Manual. Verified board console is available.

Review URL: http://codereview.chromium.org/6537006